feat(frontend-area-creation): automatically open the creation dialog if `oauth_success` is in search params
If this parameter is present, it means that the page is acceded after a redirection from the OAuth callback. The search params contain the status of the OAuth connection, the OAuth service and the id of the OAuth.
Therefore, the creation dialog could be opened so the user can continue to create its AREA.
Also add an `onOpenChange` prop to `AreaCreationDialog` so the `oauthResult` is clear when the dialog is closed: it avoids that result is displayed in the wrong place (e.g. reaction instead of action).