feat(frontend-area-creation): finish OAuth for Actions and REActions
Add an `AREActionSelection` component that permits to select a (RE)Action, and select an existant OAuth (or "create" a new connecting to the service).
The selected (RE)Action and OAuth are preserved (in session storage) during OAuth flow, and the success status, service and OAuth id is sent to the frontend through search params.