feat(frontend): add a server route for OAuth so the mobile can connect to an OAuth service using the frontend
Now, the frontend dashboard must make a request to `/oauth/[service]` with a `scope` search params.
For the mobile, it must make a request at the same route with in the search params, in addition to the scope, an access token and a redirect URI. The access token and the redirect URI are stored as cookies so the callback can access them and it knows it's the mobile that made the request.
Also, the `redirect_uri` sent to the backend is now handled server-side.